refactor: move control-common select types up from profile package
IControlCommonSelectControlById and AbstractControlCommonSelectControlById
are no longer profile-specific (they cover control-common-defined types
shared between profile and mapping-common). Move them up from
dev.metaschema.oscal.lib.model.control.profile to
dev.metaschema.oscal.lib.model.control so the package placement matches
the naming.

Updates the binding's extend-base-class reference and all consumer
imports.
