+++ src/main/java/dev/metaschema/oscal/lib/model/control/AbstractControlCommonSelectControlById.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: CC0-1.0
  */
 
-package dev.metaschema.oscal.lib.model.control.profile;
+package dev.metaschema.oscal.lib.model.control;
 
 import java.util.Collection;
 import java.util.LinkedList;
