+++ src/main/java/dev/metaschema/oscal/lib/model/control/IControlCommonSelectControlById.java
@@ -3,12 +3,11 @@
  * SPDX-License-Identifier: CC0-1.0
  */
 
-package dev.metaschema.oscal.lib.model.control.profile;
+package dev.metaschema.oscal.lib.model.control;
 
 import java.util.List;
 
 import dev.metaschema.oscal.lib.model.ProfileMatching;
-import dev.metaschema.oscal.lib.model.control.IControlSelection;
 
 public interface IControlCommonSelectControlById extends IControlSelection {
