+++ src/main/java/dev/metaschema/oscal/lib/profile/resolver/selection/IControlFilter.java
@@ -12,8 +12,8 @@ import java.util.List;
 import dev.metaschema.core.util.ObjectUtils;
 import dev.metaschema.oscal.lib.model.IncludeAll;
 import dev.metaschema.oscal.lib.model.ProfileImport;
+import dev.metaschema.oscal.lib.model.control.IControlCommonSelectControlById;
 import dev.metaschema.oscal.lib.model.control.catalog.IControl;
-import dev.metaschema.oscal.lib.model.control.profile.IControlCommonSelectControlById;
 import edu.umd.cs.findbugs.annotations.NonNull;
 
 public interface IControlFilter {
